Move logic out into modules
