+++ src/mailer.js
@@ -1,0 +1,14 @@
+const nodemailer = require('nodemailer');
+
+
+const transporter = nodemailer.createTransport({
+	host: process.env.EMAIL_SERVER,
+	port: 587,
+	secure: false, // upgrade later with STARTTLS
+	auth: {
+		user: process.env.EMAIL_USERNAME,
+		pass: process.env.EMAIL_PASSWORD
+	}
+});
+
+module.exports = transporter;
