+++ src/modules/vote/model.js
